move game to models directory
